refactor vendorsaside
